Update linux-prereqs.sh (#3585)
On Arch Linux, signing in with a GitHub or Microsoft account requires the xdg-open command which is a part of the xdg-utils package. xdg-utils is a transitive dependency of many desktop environments and the chrome browser, but I did not have it installed, so I was unable to sign in.